Refactor the main function to place file sending and receving funcitonality in a new FileTransmitter class.
