+++ ReliableUDP/ReliableUDP.cpp
@@ -10,9 +10,10 @@
 #include <vector>
 #include <chrono>
 #include <numeric>
+#include <filesystem>
 #include "Net.h"
-#include "CRC.h"
 
+#include "FileTransmitter.h"
 
 using namespace std;
 using namespace net;
@@ -23,24 +24,8 @@ const int ProtocolId = 0x11223344;
 const float DeltaTime = 1.0f / 30.0f;
 const float SendRate = 1.0f / 30.0f;
 const float TimeOut = 10.0f;
-const int PacketSize = 256;
-const int MaxFileNameLength = 128;
-const int FileDataChunkSize = PacketSize - sizeof(uint32_t);
+
 const float AckWaitTime = 2.0f;  // Time to wait for final acks
-
-#pragma pack(push, 1)
-struct FileMetadata {
-	char fileName[MaxFileNameLength];
-	uint32_t fileSize;
-	uint32_t totalChunks;
-	uint32_t crc32;
-};
-
-struct FileChunk {
-	uint32_t chunkIndex;
-	unsigned char data[FileDataChunkSize];
-};
-#pragma pack(pop)
 
 class FlowControl
 {
@@ -148,7 +133,7 @@ int main(int argc, char* argv[])
 	Mode mode = Server;
 	Address address;
 	std::string filePath;
-
+	// parse command line
 	if (argc >= 2)
 	{
 		int a, b, c, d;
@@ -156,9 +141,23 @@ int main(int argc, char* argv[])
 		{
 			mode = Client;
 			address = Address(a, b, c, d, ServerPort);
+			// fetch file path if provided
 			if (argc >= 3) {
 				filePath = argv[2];
+
+				// TODO: Validate the filepath
+				if (!filesystem::exists(filePath) && std::filesystem::is_regular_file(filePath))
+				{
+					printf("Specified file doesn't exist.\n");
+					return 1;
+				}
 			}
+		}
+		else
+		{
+			printf("client mode usage:\n"
+			"%s <ip> [file] \n", argv[0]);
+			return 1;
 		}
 	}
 
@@ -223,151 +222,52 @@ int main(int argc, char* argv[])
 
 		// send and receive packets
 
+		sendAccumulator += DeltaTime;
+		//
+
 		if (mode == Client && !filePath.empty()) {
 			static bool fileSent = false;
+
+			// TODO: instance a file transmitter
+
 			if (!fileSent) {
 				auto startTime = std::chrono::high_resolution_clock::now();
-
-				// Read file contents
-				std::ifstream ifs(filePath, std::ios::binary | std::ios::ate);
-				if (!ifs.is_open()) {
-					printf("Error opening file: %s\n", filePath.c_str());
-					return 1;
-				}
-
-				uint32_t fileSize = ifs.tellg();
-				ifs.seekg(0);
-				std::vector<unsigned char> fileData(fileSize);
-				ifs.read(reinterpret_cast<char*>(fileData.data()), fileSize);
-				ifs.close();
-
-				// Calculate CRC32
-				uint32_t fileCRC = CRC::Calculate(fileData.data(), fileData.size(), CRC::CRC_32());
-
-				// Prepare metadata
-				FileMetadata metadata;
-				const char* filename = strrchr(filePath.c_str(), '\\');
-				filename = filename ? filename + 1 : filePath.c_str();
-				strncpy(metadata.fileName, filename, MaxFileNameLength - 1);
-				metadata.fileName[MaxFileNameLength - 1] = '\0';
-				metadata.fileSize = fileSize;
-				metadata.totalChunks = (fileSize + FileDataChunkSize - 1) / FileDataChunkSize;
-				metadata.crc32 = fileCRC;
-
-				// Send metadata
-				connection.SendPacket(reinterpret_cast<unsigned char*>(&metadata), sizeof(metadata));
-
-				// Splits file into chunks and sends them
-				for (uint32_t i = 0; i < metadata.totalChunks; i++) {
-					FileChunk chunk;
-					chunk.chunkIndex = i;
-					size_t offset = i * FileDataChunkSize;
-					size_t chunkSize = (((FileDataChunkSize) < (fileSize - offset)) ? (FileDataChunkSize) : (fileSize - offset))
-						;
-
-					// Copy from memory buffer instead of re-reading file
-					memcpy(chunk.data, fileData.data() + offset, chunkSize);
-
-					// Fill remaining space with zeros if needed
-					if (chunkSize < FileDataChunkSize) {
-						memset(chunk.data + chunkSize, 0, FileDataChunkSize - chunkSize);
-					}
-
-					connection.SendPacket(reinterpret_cast<unsigned char*>(&chunk), sizeof(chunk));
-				}
-
-				// Wait for all acks with timeout
-				float ackWaitTime = 0.0f;
-				while (ackWaitTime < AckWaitTime &&
-					connection.GetReliabilitySystem().GetAckedPackets() < metadata.totalChunks + 1) {
-					connection.Update(DeltaTime);
-					ackWaitTime += DeltaTime;
-					net::wait(DeltaTime);
-				}
-
-				// Calculate actual transfer time
-				auto endTime = std::chrono::high_resolution_clock::now();
-				float transferTime = std::chrono::duration<float>(endTime - startTime).count();
-
-				// Calculate speed in Mbps (1 megabit = 1,000,000 bits)
-				float fileSizeBits = fileSize * 8.0f;
-				float speedMbps = (fileSizeBits / transferTime) / 1000000.0f;
-
-				printf("Transfer complete!\n");
-				printf("File size: %.2f KB\n", fileSize / 1024.0f);
-				printf("Transfer time: %.2fs\n", transferTime);
-				printf("Effective speed: %.2f Mbps\n", speedMbps);
-
-				printf("Calculated File CRC: 0x%08X\n", fileCRC);
-
-				fileSent = true;
 			}
 		}
 
-		static std::ofstream outputFile;
-		static std::vector<unsigned char> receivedFile;
-		static FileMetadata receivedMetadata;
-		static uint32_t receivedChunks = 0;
+		// send packets at a fixed rate
+		while (sendAccumulator > 1.0f / sendRate)
+		{
+			// invoke the method in the file transmitter
+			// send file chunks sequentially
+			// stop sending when all chunks have been sent and no more data is available
+			unsigned char packet[PacketSize];
+			memset(packet, 0, sizeof(packet));
+			// In the CLIENT mode
+			// 
+			// 1. Specify the type of the packet (Message ID) and contents
+			// 2. Contents for the packets containing the file meta data
+			// 		 get the seralized meta data of the file from the file transmitter and put it into the packets to send
+			// 3. contents for the packets containing the file chunks
+			//  get the file chunks from the file transmitter and put them into the packet to send
+			// 4. contents for the packets indicating the end of the file
+			snprintf((char*)packet, sizeof(packet), "Hello World <<%d>>", sentCount);
+			connection.SendPacket(packet, sizeof(packet));
+			++sentCount;
+			sendAccumulator -= 1.0f / sendRate;
+		}
 
-		while (true) {
+		while (true)
+		{
+			// in the server mode, receive packets of the file 
+			// invoke methods in the filetransmitter to save the file back to listening state after having verified the file
 			unsigned char packet[PacketSize];
 			int bytes_read = connection.ReceivePacket(packet, sizeof(packet));
-			if (bytes_read == 0) break;
-
-			if (bytes_read == sizeof(FileMetadata)) {
-				// Received metadata
-				memcpy(&receivedMetadata, packet, sizeof(receivedMetadata));
-				receivedFile.resize(receivedMetadata.fileSize);
-				printf("Receiving file: %s (%u bytes)\n",
-					receivedMetadata.fileName, receivedMetadata.fileSize);
-				printf("Expecting %u bytes in %u chunks\n",
-					receivedMetadata.fileSize, receivedMetadata.totalChunks);
-				receivedFile.reserve(receivedMetadata.fileSize); // Pre-allocate exact size
-			}
-			else if (bytes_read == sizeof(FileChunk)) {
-				// Received file chunk
-				FileChunk chunk;
-				memcpy(&chunk, packet, sizeof(FileChunk));
-
-				// Ensure vector is properly sized before writing
-				size_t offset = chunk.chunkIndex * FileDataChunkSize;
-				size_t remaining = receivedMetadata.fileSize - offset;
-				size_t copySize = (((FileDataChunkSize) < (remaining)) ? (FileDataChunkSize) : (remaining))
-					;
-
-				if (offset + copySize > receivedFile.size()) {
-					receivedFile.resize(offset + copySize);
-				}
-
-				memcpy(receivedFile.data() + offset, chunk.data, copySize);
-				receivedChunks++;
-
-				// Check if transfer complete
-				if (receivedChunks == receivedMetadata.totalChunks) {
-					// Verify CRC
-					printf("Finalizing transfer with %u/%u chunks received\n",
-						receivedChunks, receivedMetadata.totalChunks);
-					printf("Received file size: %zu bytes\n", receivedFile.size());
-					printf("Original CRC claim: 0x%08X\n", receivedMetadata.crc32);
-
-					uint32_t calculatedCRC = CRC::Calculate(receivedFile.data(),
-						receivedFile.size(), CRC::CRC_32());
-					printf("Server Calculated CRC: 0x%08X\n", calculatedCRC);
-
-					if (calculatedCRC == receivedMetadata.crc32) {
-						std::ofstream outFile(receivedMetadata.fileName, std::ios::binary);
-						outFile.write(reinterpret_cast<char*>(receivedFile.data()),
-							receivedFile.size());
-						printf("File received and verified successfully!\n");
-					}
-					else {
-						printf("ERROR: File verification failed!\n");
-					}
-				}
-
-				printf("Received chunk %u/%u (%zu bytes)\n",
-					chunk.chunkIndex + 1, receivedMetadata.totalChunks, copySize);
-			}
+			if (bytes_read == 0)
+				break;
+			printf("%s\n", packet);
+			// verify the file after receiving the EOF packet
+			// save the file and go back to listening state if the file integrity is verified
 		}
 
 
@@ -388,7 +288,9 @@ int main(int argc, char* argv[])
 		// update connection
 
 		connection.Update(DeltaTime);
-
+		// TODO: 
+		// pasre received message.
+		
 		// show connection stats
 
 		statsAccumulator += DeltaTime;
